+++ src/main/java/org/betonquest/betonquest/conversation/ConversationData.java
@@ -1,12 +1,16 @@
 package org.betonquest.betonquest.conversation;
 
+import net.kyori.adventure.text.Component;
+import net.kyori.adventure.text.serializer.legacy.LegacyComponentSerializer;
 import org.apache.commons.lang3.StringUtils;
 import org.betonquest.betonquest.api.config.quest.QuestPackage;
 import org.betonquest.betonquest.api.feature.FeatureAPI;
 import org.betonquest.betonquest.api.logger.BetonQuestLogger;
+import org.betonquest.betonquest.api.message.MessageParser;
 import org.betonquest.betonquest.api.profile.Profile;
 import org.betonquest.betonquest.api.quest.QuestException;
 import org.betonquest.betonquest.api.quest.QuestTypeAPI;
+import org.betonquest.betonquest.data.PlayerDataStorage;
 import org.betonquest.betonquest.id.ConditionID;
 import org.betonquest.betonquest.id.ConversationID;
 import org.betonquest.betonquest.id.EventID;
@@ -16,7 +20,6 @@ import org.betonquest.betonquest.kernel.processor.quest.VariableProcessor;
 import org.betonquest.betonquest.message.ParsedSectionMessage;
 import org.betonquest.betonquest.util.Utils;
 import org.betonquest.betonquest.variables.GlobalVariableResolver;
-import org.bukkit.ChatColor;
 import org.bukkit.configuration.ConfigurationSection;
 import org.jetbrains.annotations.Nullable;
 
@@ -62,6 +65,16 @@ public class ConversationData {
     private final VariableProcessor variableProcessor;
 
     /**
+     * Message parser to parse messages.
+     */
+    private final MessageParser messageParser;
+
+    /**
+     * Player data storage to get the player language.
+     */
+    private final PlayerDataStorage playerDataStorage;
+
+    /**
      * The {@link QuestPackage} this conversation is in.
      */
     private final QuestPackage pack;
@@ -100,18 +113,24 @@ public class ConversationData {
      * @param questTypeAPI      the quest type api
      * @param featureAPI        the feature api
      * @param variableProcessor the variable processor to create new variables
+     * @param messageParser     the message parser to parse messages
+     * @param playerDataStorage the player data storage to get the player language
      * @param pack              the package of the conversation this data represents
      * @param convSection       the configuration section of the conversation
      * @param publicData        the external used data
      * @throws QuestException when there is a syntax error in the defined conversation or
      *                        when conversation options cannot be resolved or {@code convSection} is null
      */
-    public ConversationData(final BetonQuestLogger log, final QuestTypeAPI questTypeAPI, final FeatureAPI featureAPI, final VariableProcessor variableProcessor,
-                            final QuestPackage pack, final ConfigurationSection convSection, final PublicData publicData) throws QuestException {
+    public ConversationData(final BetonQuestLogger log, final QuestTypeAPI questTypeAPI, final FeatureAPI featureAPI,
+                            final VariableProcessor variableProcessor, final MessageParser messageParser,
+                            final PlayerDataStorage playerDataStorage, final QuestPackage pack,
+                            final ConfigurationSection convSection, final PublicData publicData) throws QuestException {
         this.log = log;
         this.questTypeAPI = questTypeAPI;
         this.featureAPI = featureAPI;
         this.variableProcessor = variableProcessor;
+        this.messageParser = messageParser;
+        this.playerDataStorage = playerDataStorage;
         this.pack = pack;
         this.convName = publicData.convName();
         this.publicData = publicData;
@@ -349,12 +368,11 @@ public class ConversationData {
      * Respects extended options.
      *
      * @param profile the profile of the player
-     * @param lang    the desired language of the text
      * @param option  the option
      * @return the text of the specified option in the specified language
      */
     @Nullable
-    public String getText(@Nullable final Profile profile, final String lang, final ResolvedOption option) {
+    public String getText(@Nullable final Profile profile, final ResolvedOption option) {
         final ConversationOption opt;
         if (option.type() == NPC) {
             opt = option.conversationData().npcOptions.get(option.name());
@@ -364,7 +382,7 @@ public class ConversationData {
         if (opt == null) {
             return null;
         }
-        return opt.getText(profile, lang);
+        return opt.getText(profile);
     }
 
     /**
@@ -515,16 +533,15 @@ public class ConversationData {
          * Returns "Quester" in case of an exception.
          *
          * @param log     the logger used when the name could not be resolved
-         * @param lang    language key
          * @param profile the profile to resolve the variable
          * @return the quester's name in the specified language
          */
-        public String getQuester(final BetonQuestLogger log, final String lang, final Profile profile) {
+        public Component getQuester(final BetonQuestLogger log, final Profile profile) {
             try {
-                return ChatColor.translateAlternateColorCodes('&', quester.getResolved(lang, profile));
+                return quester.asComponent(profile);
             } catch (final QuestException e) {
                 log.warn("Could not get Quester's name! Using 'Quester' instead, reason: " + e.getMessage(), e);
-                return "Quester";
+                return Component.text("Quester");
             }
         }
     }
@@ -632,7 +649,7 @@ public class ConversationData {
             }
             final ParsedSectionMessage text;
             try {
-                text = new ParsedSectionMessage(variableProcessor, pack, conv, "text");
+                text = new ParsedSectionMessage(variableProcessor, messageParser, playerDataStorage, pack, conv, "text");
             } catch (final QuestException e) {
                 throw new QuestException("Could not load text for " + optionName + " " + type.getReadable() + ": " + e.getMessage(), e);
             }
@@ -652,26 +669,25 @@ public class ConversationData {
          * Returns the text of this option in the given language.
          *
          * @param profile the profile of the player to get the text for
-         * @param lang    the language to get the text in
          * @return the text of this option in the given language
          */
-        public String getText(@Nullable final Profile profile, final String lang) {
-            return getText(profile, lang, new ArrayList<>());
+        public String getText(@Nullable final Profile profile) {
+            return getText(profile, new ArrayList<>());
         }
 
-        private String getText(@Nullable final Profile profile, final String lang, final List<String> optionPath) {
+        private String getText(@Nullable final Profile profile, final List<String> optionPath) {
             // Prevent infinite loops
             if (optionPath.contains(getName())) {
                 return "";
             }
             optionPath.add(getName());
 
-            final StringBuilder text = new StringBuilder(getText(lang, profile));
+            final StringBuilder text = new StringBuilder(getFormattedText(profile));
 
             if (profile != null) {
                 for (final String extend : extendLinks) {
                     if (questTypeAPI.conditions(profile, getOption(extend, type).getConditions())) {
-                        text.append(getOption(extend, type).getText(profile, lang, optionPath));
+                        text.append(getOption(extend, type).getText(profile, optionPath));
                         break;
                     }
                 }
@@ -680,13 +696,13 @@ public class ConversationData {
             return text.toString();
         }
 
-        private String getText(final String lang, @Nullable final Profile profile) {
+        private String getFormattedText(@Nullable final Profile profile) {
             if (text == null) {
                 log.warn(pack, "No text in conversation '" + convName + "'!");
                 return "";
             }
             try {
-                return Utils.format(text.getResolved(lang, profile));
+                return Utils.format(LegacyComponentSerializer.legacySection().serialize(text.asComponent(profile)));
             } catch (final QuestException e) {
                 log.warn(pack, "Could not resolve message in conversation '" + convName + "': " + e.getMessage(), e);
                 return "";