+++ src/main/java/org/betonquest/betonquest/feature/Backpack.java
@@ -1,5 +1,6 @@
 package org.betonquest.betonquest.feature;
 
+import net.kyori.adventure.text.Component;
 import org.apache.commons.lang3.tuple.Pair;
 import org.betonquest.betonquest.BetonQuest;
 import org.betonquest.betonquest.api.bukkit.event.QuestCompassTargetChangeEvent;
@@ -520,15 +521,15 @@ public class Backpack implements Listener {
                     log.warn("Could not find item: " + e.getMessage(), e);
                     compass = new ItemStack(Material.COMPASS);
                 }
-                final String name;
+                final Component name;
                 try {
-                    name = comp.names().getResolved(lang, onlineProfile);
+                    name = comp.names().asComponent(onlineProfile);
                 } catch (final QuestException e) {
                     log.warn("Could not get name for compass '" + entry.getKey() + "', skipping it: " + e.getMessage(), e);
                     continue;
                 }
                 final ItemMeta meta = compass.getItemMeta();
-                meta.setDisplayName(name.replace("_", " ").replace("&", "§"));
+                meta.displayName(name);
                 compass.setItemMeta(meta);
                 content[index] = compass;
                 index++;