+++ src/main/java/org/betonquest/betonquest/kernel/processor/feature/CompassProcessor.java
@@ -2,7 +2,9 @@ package org.betonquest.betonquest.kernel.processor.feature;
 
 import org.betonquest.betonquest.api.config.quest.QuestPackage;
 import org.betonquest.betonquest.api.logger.BetonQuestLogger;
+import org.betonquest.betonquest.api.message.MessageParser;
 import org.betonquest.betonquest.api.quest.QuestException;
+import org.betonquest.betonquest.data.PlayerDataStorage;
 import org.betonquest.betonquest.feature.QuestCompass;
 import org.betonquest.betonquest.id.CompassID;
 import org.betonquest.betonquest.id.ItemID;
@@ -24,19 +26,34 @@ public class CompassProcessor extends SectionProcessor<CompassID, QuestCompass> 
     private final VariableProcessor variableProcessor;
 
     /**
+     * Message parser to parse messages.
+     */
+    private final MessageParser messageParser;
+
+    /**
+     * Player data storage to get the player language.
+     */
+    private final PlayerDataStorage playerDataStorage;
+
+    /**
      * Create a new QuestProcessor to store {@link QuestCompass}es.
      *
      * @param log               the custom logger for this class
      * @param variableProcessor the variable processor to create new variables
+     * @param messageParser     the message parser to parse messages
+     * @param playerDataStorage the player data storage to get the player language
      */
-    public CompassProcessor(final BetonQuestLogger log, final VariableProcessor variableProcessor) {
+    public CompassProcessor(final BetonQuestLogger log, final VariableProcessor variableProcessor,
+                            final MessageParser messageParser, final PlayerDataStorage playerDataStorage) {
         super(log, "Compass", "compass");
         this.variableProcessor = variableProcessor;
+        this.messageParser = messageParser;
+        this.playerDataStorage = playerDataStorage;
     }
 
     @Override
     protected QuestCompass loadSection(final QuestPackage pack, final ConfigurationSection section) throws QuestException {
-        final ParsedSectionMessage names = new ParsedSectionMessage(variableProcessor, pack, section, "name");
+        final ParsedSectionMessage names = new ParsedSectionMessage(variableProcessor, messageParser, playerDataStorage, pack, section, "name");
         final String location = section.getString("location");
         if (location == null) {
             throw new QuestException("Location not defined");