+++ src/main/java/org/betonquest/betonquest/message/ParsedSectionMessage.java
@@ -1,14 +1,14 @@
 package org.betonquest.betonquest.message;
 
 import org.betonquest.betonquest.api.config.quest.QuestPackage;
-import org.betonquest.betonquest.api.profile.Profile;
+import org.betonquest.betonquest.api.message.MessageParser;
 import org.betonquest.betonquest.api.quest.QuestException;
 import org.betonquest.betonquest.config.Config;
+import org.betonquest.betonquest.data.PlayerDataStorage;
 import org.betonquest.betonquest.instruction.variable.VariableString;
 import org.betonquest.betonquest.kernel.processor.quest.VariableProcessor;
 import org.betonquest.betonquest.variables.GlobalVariableResolver;
 import org.bukkit.configuration.ConfigurationSection;
-import org.jetbrains.annotations.Nullable;
 
 import java.util.HashMap;
 import java.util.Map;
@@ -16,48 +16,44 @@ import java.util.Map;
 /**
  * Messages loaded from a Configuration Section identified by languages.
  */
-public class ParsedSectionMessage {
-    /**
-     * The default language.
-     */
-    private final String defaultLanguage = Config.getLanguage();
-
-    /**
-     * Maps the language to the message.
-     */
-    private final Map<String, VariableString> messages;
+public class ParsedSectionMessage extends ParsedMessage {
 
     /**
      * Loads value(s) from a key in a section, potentially identified by a language key.
-     * When there is no section the value will be identified by the default language.
+     * When there is no section, the value will be identified by the default language.
      *
      * @param variableProcessor the variable processor to create new variables
+     * @param messageParser     the message parser to parse the message
+     * @param playerDataStorage the player data storage to get the player's language
      * @param pack              the pack to resolve variables
      * @param section           the section to load from
      * @param path              where the value(s) are stored in the section
      * @throws QuestException if there is no value, the default language is missing or the section format is invalid
      */
-    public ParsedSectionMessage(final VariableProcessor variableProcessor, final QuestPackage pack,
+    public ParsedSectionMessage(final VariableProcessor variableProcessor, final MessageParser messageParser,
+                                final PlayerDataStorage playerDataStorage, final QuestPackage pack,
                                 final ConfigurationSection section, final String path) throws QuestException {
+        super(messageParser, parse(variableProcessor, pack, section, path), playerDataStorage);
+    }
+
+    private static Map<String, VariableString> parse(final VariableProcessor variableProcessor, final QuestPackage pack,
+                                                     final ConfigurationSection section, final String path) throws QuestException {
         if (section.isConfigurationSection(path)) {
-            messages = parseSection(variableProcessor, pack, section, path);
-            if (!messages.containsKey(defaultLanguage)) {
-                throw new QuestException("No message for default language!");
-            }
+            return parseSection(variableProcessor, pack, section, path);
         } else if (section.isString(path)) {
             final String raw = GlobalVariableResolver.resolve(pack, section.getString(path));
             if (raw == null) {
                 throw new QuestException("No string value for '" + path + "'!");
             }
-            messages = Map.of(defaultLanguage, new VariableString(variableProcessor, pack, raw));
+            return Map.of(Config.getLanguage(), new VariableString(variableProcessor, pack, raw));
         } else {
             throw new QuestException("The '" + path + "' is missing!");
         }
     }
 
-    private Map<String, VariableString> parseSection(final VariableProcessor variableProcessor, final QuestPackage pack,
-                                                     final ConfigurationSection section, final String path) throws QuestException {
-        final ConfigurationSection subSection = section.getConfigurationSection(path);
+    private static Map<String, VariableString> parseSection(final VariableProcessor variableProcessor, final QuestPackage pack,
+                                                            final ConfigurationSection messageSection, final String path) throws QuestException {
+        final ConfigurationSection subSection = messageSection.getConfigurationSection(path);
         if (subSection == null) {
             throw new QuestException("No configuration section for '" + path + "'!");
         }
@@ -73,30 +69,5 @@ public class ParsedSectionMessage {
             throw new QuestException("No values defined!");
         }
         return messages;
-    }
-
-    /**
-     * Resolves the message in the language or default language.
-     *
-     * @param language the preferred language to get the message
-     * @param profile  the profile to resolve the variable
-     * @return the resolved message for the language
-     * @throws IllegalStateException when there is no message for the requested language and default language
-     * @throws QuestException        or the variable can't be resolved
-     */
-    public String getResolved(final String language, @Nullable final Profile profile) throws QuestException {
-        return get(language).getValue(profile);
-    }
-
-    private VariableString get(final String language) {
-        VariableString name = messages.get(language);
-        if (name != null) {
-            return name;
-        }
-        name = messages.get(defaultLanguage);
-        if (name != null) {
-            return name;
-        }
-        throw new IllegalStateException("No text for language '" + language + "' or default!");
     }
 }